VisaFlow/requirements.txt
Django>=3.2
reportlab
djangorestframework
drf-yasg